+++ src/LightsprintGL/Integration/SceneViewer.cpp
@@ -39,7 +39,7 @@ void error(const char* message, bool gfxRelated)
 // globals are ugly, but required by GLUT design with callbacks
 
 class Solver*              solver = NULL;
-Camera                     eye(-1.856,1.440,2.097,2.404,0.000,0.020,1.3,90.0,0.1,1000.0);
+Camera                     eye(-1.856f,1.440f,2.097f,2.404f,0,0.02f,1.3f,90,0.1f,1000);
 enum SelectionType {ST_CAMERA, ST_LIGHT, ST_OBJECT};
 SelectionType              selectedType = ST_CAMERA;
 unsigned                   selectedLightIndex = 0; // index into lights, light controlled by mouse/arrows
@@ -210,10 +210,10 @@ void keyboard(unsigned char c, int x, int y)
 {
 	switch (c)
 	{
-		case '+': brightness *= 1.2; break;
-		case '-': brightness /= 1.2; break;
-		case '*': gamma *= 1.2; break;
-		case '/': gamma /= 1.2; break;
+		case '+': brightness *= 1.2f; break;
+		case '-': brightness /= 1.2f; break;
+		case '*': gamma *= 1.2f; break;
+		case '/': gamma /= 1.2f; break;
 
 		case 'a':
 		case 'A': speedLeft = 1; break;
@@ -267,11 +267,11 @@ void reshape(int w, int h)
 	winWidth = w;
 	winHeight = h;
 	glViewport(0, 0, w, h);
-	eye.aspect = (double) winWidth / (double) winHeight;
+	eye.aspect = winWidth/(float)winHeight;
 	Texture* texture = Texture::createShadowmap(64,64);
 	GLint shadowDepthBits = texture->getTexelBits();
 	delete texture;
-	glPolygonOffset(1, 12 << (shadowDepthBits-16) );
+	glPolygonOffset(1,(GLfloat)(12<<(shadowDepthBits-16)));
 }
 
 void mouse(int button, int state, int x, int y)
@@ -302,17 +302,17 @@ void passive(int x, int y)
 	{
 		if(selectedType==ST_CAMERA || selectedType==ST_OBJECT)
 		{
-			eye.angle -= 0.005*x;
-			eye.angleX -= 0.005*y;
-			CLAMP(eye.angleX,-M_PI*0.49f,M_PI*0.49f);
+			eye.angle -= 0.005f*x;
+			eye.angleX -= 0.005f*y;
+			CLAMP(eye.angleX,(float)(-M_PI*0.49),(float)(M_PI*0.49));
 		}
 		else
 		if(selectedType==ST_LIGHT)
 		{
 			Camera* light = solver->realtimeLights[selectedLightIndex]->getParent();
-			light->angle -= 0.005*x;
-			light->angleX -= 0.005*y;
-			CLAMP(light->angleX,-M_PI*0.49f,M_PI*0.49f);
+			light->angle -= 0.005f*x;
+			light->angleX -= 0.005f*y;
+			CLAMP(light->angleX,(float)(-M_PI*0.49),(float)(M_PI*0.49));
 			solver->reportDirectIlluminationChange(true);
 			solver->realtimeLights[selectedLightIndex]->dirty = true;
 			// changes position a bit, together with rotation
